refactoring the folders
