+++ user_analytics.py
